pcie_driver: Fix compilation warning.

+++ drivers/pcie/src/kmem.c
@@ -306,7 +306,7 @@ int pcidriver_mmap_kmem(pcidriver_privdata_t *privdata, struct vm_area_struct *v
 #endif
 
 	mod_info_dbg("Mapping address %08lx / PFN %08lx\n",
-			virt_to_phys((void*)kmem_entry->cpua),
+			(long unsigned int)virt_to_phys((void*)kmem_entry->cpua),
 			page_to_pfn(virt_to_page((void*)kmem_entry->cpua)));
 
 	ret = remap_pfn_range_cpua_compat(
